move api.js
